ADD Tip Router tip 완료되면 main page로 이동
