+++ frontend/src/api/file/index.ts
@@ -1,0 +1,11 @@
+import { fileInstacne } from "../index";
+import { ImgRespData, ImgDataResp } from "@/api/file/type";
+
+export function submitImg(img: FormData): Promise<ImgDataResp> {
+  return fileInstacne.post("", img, {
+    headers: {
+      "Content-Type": "multipart/form-data"
+    },
+    withCredentials: true
+  });
+}
