+++ frontend/src/api/file/type.ts
@@ -1,0 +1,15 @@
+export interface ImgData {
+  file: File;
+  user: number;
+  id?: number;
+}
+
+export interface ImgRespData {
+  file: string;
+  user: number;
+  id: number;
+}
+
+export interface ImgDataResp {
+  data: ImgRespData;
+}
