+++ frontend/src/api/tip/type.ts
@@ -27,22 +27,6 @@ export interface TipData {
   hashtags: HashTag[];
 }
 
-export interface ImgData {
-  file: File;
-  user: number;
-  id?: number;
-}
-
-export interface ImgRespData {
-  file: string;
-  user: number;
-  id: number;
-}
-
-export interface ImgDataResp extends AxiosResponse {
-  data: ImgRespData;
-}
-
 export interface TipDataResp extends AxiosResponse {
   data: Tip;
 }
